+++ lib/exograph/storage/file_record.ex
@@ -13,6 +13,7 @@ defmodule Exograph.Storage.FileRecord do
     field(:package_version_id, :integer)
     field(:path, :string)
     field(:source, :string)
+    field(:ast, :binary)
     field(:comments_text, :string)
     field(:sha256, :string)
 
@@ -25,8 +26,12 @@ defmodule Exograph.Storage.FileRecord do
       package_version_id: file.package_version_id,
       path: file.path,
       source: file.source,
+      ast: compressed_binary(file.ast),
       comments_text: file.comments_text,
       sha256: file.sha256
     }
   end
+
+  defp compressed_binary(nil), do: nil
+  defp compressed_binary(term), do: :erlang.term_to_binary(term, [:compressed])
 end