+++ lib/exograph/storage/fragment_record.ex
@@ -19,7 +19,9 @@ defmodule Exograph.Storage.FragmentRecord do
     field(:file, :string, virtual: true)
     field(:source, :string, virtual: true)
     field(:content_hash, :binary)
-    field(:ast, :binary)
+    field(:ast, :binary, virtual: true)
+    field(:node_pre, :integer)
+    field(:node_post, :integer)
 
     field(:kind, Ecto.Enum,
       values: [:unknown, :module, :expression, :def, :defp, :defmacro, :defmacrop]
@@ -32,7 +34,7 @@ defmodule Exograph.Storage.FragmentRecord do
     field(:end_line, :integer)
     field(:mass, :integer)
     field(:exact_hash, :binary)
-    field(:terms, {:array, :integer}, default: [])
+    field(:terms, {:array, :integer}, default: [], virtual: true)
     field(:sub_hashes, {:array, :integer}, default: [])
 
     timestamps(type: :utc_datetime_usec)
@@ -41,7 +43,7 @@ defmodule Exograph.Storage.FragmentRecord do
   def changeset(record, attrs) do
     record
     |> cast(attrs, __schema__(:fields))
-    |> validate_required([:ast, :kind, :line, :mass])
+    |> validate_required([:kind, :line, :mass])
   end
 
   def from_fragment(%Fragment{} = fragment) do
@@ -50,7 +52,8 @@ defmodule Exograph.Storage.FragmentRecord do
       package_version_id: fragment.package_version_id,
       file_id: fragment.file_id,
       content_hash: fragment.content_hash,
-      ast: compressed_binary(fragment.ast),
+      node_pre: fragment.node_pre,
+      node_post: fragment.node_post,
       kind: fragment.kind,
       module: fragment.module,
       name: fragment.name,
@@ -59,12 +62,13 @@ defmodule Exograph.Storage.FragmentRecord do
       end_line: fragment.end_line,
       mass: fragment.mass,
       exact_hash: encode_hash(fragment.exact_hash),
-      terms: MapSet.to_list(fragment.terms),
       sub_hashes: MapSet.to_list(fragment.sub_hashes)
     }
   end
 
-  def to_fragment(%__MODULE__{} = record) do
+  def to_fragment(%__MODULE__{} = record, file_ast \\ nil) do
+    ast = record.ast || Exograph.AST.Locator.slice(file_ast, record.node_pre, record.node_post)
+
     %Fragment{
       id: record.id,
       package_id: record.package_id,
@@ -74,7 +78,9 @@ defmodule Exograph.Storage.FragmentRecord do
       file: record.file,
       source: record.source,
       content_hash: record.content_hash,
-      ast: :erlang.binary_to_term(record.ast, [:safe]),
+      ast: ast,
+      node_pre: record.node_pre,
+      node_post: record.node_post,
       kind: record.kind,
       module: record.module,
       name: record.name,
@@ -99,7 +105,6 @@ defmodule Exograph.Storage.FragmentRecord do
 
   defp encode_hash(hash), do: to_string(hash)
 
-  defp compressed_binary(term), do: :erlang.term_to_binary(term, [:compressed])
   defp mapset(nil), do: MapSet.new()
   defp mapset(values), do: MapSet.new(values)
 end